+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 CHANGELOG
 ==============
 
+1.1.0
+---------------
+ * Added yandex ruleset http://translit-online.ru/yandex.html
+ * Added custom slug rule slugify provider
+ 
 1.0.0
 ---------------
  * Ready